bug fixed

+++ util/io.h
@@ -2,7 +2,7 @@
 #ifndef _IO_H
 #define _IO_H
 
-int read_lime_gauge_field_doubleprec(float * config, char * filename,
+int read_lime_gauge_field_doubleprec(double * config, char * filename,
 				     const int T, const int LX, const int LY, const int LZ);
 
 int read_lime_gauge_field_singleprec(float * config, char * filename,
